archive/local-test/requirements.txt
# OpenAI API for transcription and article generation
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# YouTube video processing (optional, for test_script.py)
yt-dlp>=2023.12.30

# Audio processing dependencies
# Note: FFmpeg must be installed separately on your system
# macOS: brew install ffmpeg
# Ubuntu: sudo apt-get install ffmpeg
# Windows: Download from https://ffmpeg.org/download.html